Split base RPCs into separate classes
